Moving scripts outside of src dir, creating subdirectories, renaming confluence wiki to json, and creating readme for usage.
