+++ ChromaDB/scripts/processors/__init__.py
@@ -1,0 +1,11 @@
+from .base_document_processor import BaseDocumentProcessor
+from .confluence_documents import ConfluenceDocuments
+from .research_papers import ResearchPapers
+from .cambridge_bay_articles import CambridgeBayArticles
+
+__all__ = [
+    'BaseDocumentProcessor',
+    'ConfluenceDocuments',
+    'ResearchPapers',
+    'CambridgeBayArticles'
+]
