+++ ChromaDB/scripts/processors/cambridge_bay_articles.py
@@ -1,22 +1,22 @@
-from base_document_processor import BaseDocumentProcessor
+from .base_document_processor import BaseDocumentProcessor
 from typing import List, Dict
 import re
 
 
-class ResearchPapers(BaseDocumentProcessor):
+class CambridgeBayArticles(BaseDocumentProcessor):
     def __init__(self, docs: List[Dict]):
         super().__init__(docs)
         
     def clean(self) -> List[str]:
-        return [re.sub(r"\\[a-zA-Z]+", "", doc['content']) for doc in self.docs]
+        return [re.sub(r"\s+", " ", re.sub(r"<[^>]+>", "", doc['content'])).strip() for doc in self.docs]
 
     def create_metadata(self) -> List[Dict]:
         metadata = []
         for doc in self.docs:
             source = self._extract_source(doc['content'])
-            title = self._extract_title(doc['content'], source)
+            title = self._extract_title_from_url(source)
             metadata.append({
-                'source_type': 'paper',
+                'source_type': 'web_article',
                 'length': len(doc['content']),
                 'title': title,
                 'source': source,
@@ -25,13 +25,10 @@ class ResearchPapers(BaseDocumentProcessor):
             })
         return metadata
 
-    def chunk_with_metadata(self, max_tokens: int = 500, overlap: int = 50) -> List[Dict]:
+    def chunk_with_metadata(self, max_tokens: int = 400, overlap: int = 50) -> List[Dict]:
         cleaned_docs = self.clean()
         metadata = self.create_metadata()
         return self._section_chunking(cleaned_docs, metadata, max_tokens, overlap)
 
-    def _extract_title(self, text: str, source: str) -> str:
-        for line in text.splitlines():
-            if line.strip() and not line.startswith("#"):
-                return line.strip()[:100]
-        return source[:80] if source != 'Unknown' else "Untitled"
+    def _extract_title_from_url(self, url: str) -> str:
+        return url.rstrip('/').split('/')[-1] if url != 'Unknown' else 'Untitled'
